SCB-252 add extension module metrics-push
