+++ metrics/metrics-extension/metrics-push/pom.xml
@@ -20,19 +20,19 @@
   xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
   xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <parent>
-    <artifactId>metrics-integration</artifactId>
+    <artifactId>metrics-extension</artifactId>
     <groupId>org.apache.servicecomb</groupId>
     <version>0.6.0-SNAPSHOT</version>
   </parent>
   <modelVersion>4.0.0</modelVersion>
 
-  <artifactId>metrics-overwatch</artifactId>
-  <name>Java Chassis::Metrics::Integration::Overwatch</name>
+  <artifactId>metrics-push</artifactId>
+  <name>Java Chassis::Metrics::Extension::Push</name>
 
   <dependencies>
     <dependency>
       <groupId>org.apache.servicecomb</groupId>
-      <artifactId>metrics-push</artifactId>
+      <artifactId>metrics-core</artifactId>
     </dependency>
   </dependencies>
