+++ metrics/metrics-integration/metrics-overwatch/src/main/java/org/apache/servicecomb/metrics/overwatch/dto/InstanceStatus.java
@@ -15,10 +15,11 @@
  * limitations under the License.
  */
 
-package org.apache.servicecomb.metrics.overwatch;
+package org.apache.servicecomb.metrics.overwatch.dto;
 
 public class InstanceStatus {
   private final Integer rpm;
+
   private final Integer fpm;
 
   public Integer getRpm() {
